backend/requirements.txt
ultralytics==8.2.27
django==5.0.6
django-rest-framework==0.1.0
requests==2.32.3
psycopg2-binary==2.9.9
numpy<2.0.0
python-dotenv==1.0.1